feat: UCAN authorized signing server (#7)
Adds a Ucanto server at `/` for getting your messages signed. You MUST have a delegation to `pdp/sign/*` from the signing service to use.

This PR is currently backwards compatible with the existing service (I haven't removed the old routes).

It also adds an `access/grant` handler so we don't need to ask the delegator for a delegation to use this service (Note: it is a TODO to actually validate the request).

BREAKING CHANGE: The client for this service is now a Ucanto server client. Old server routes are still available temporarily.

resolves https://github.com/storacha/piri-signing-service/issues/2